Update hop_count implementation.
This makes the min / max value 1 - 8 (000 = 1, 111 = 8). While this is
not an official feature yet, this makes sure that in the future we can
use up to 8 hops. This also changes the initial hop_count to 1. Before
it was set to 0, and then in the metadata we would use hop_count + 1.