Fix mips build.
As mips is a tier-3 target, it can only be built using the nightly
toolchain. Because of the latest changes, we need to update the
nightly toolchain. However, the latest toolchain does not play nicely
with cargo-cross 0.2.5. This problem has been solved, but not yet
released which is why we install it using dev-dependencies.