Implement encryption of event / command payloads.
This implements the encryption of event and command payloads, to avoid
the exposure of possible sensitive data (e.g. configuration parameters).

This also adds a new root_key configuration, which will be used to derrive
the keys for mic calculation and encryption. In case the signing_key
configuration is present, then this key will be used for mic calculation
(for backwards compatibility).